API fix
Fixed the API to handle the case when the parsing client decides to
stop parsing in the middle of the buffer, but then wishes continue
later where it left off.  The parse function now returns the number of
bytes parsed so that the client can set its cursor in the buffer
appropriately.

Also removed a redundant {Xx}ml_ prefix from internal and some
exported symbols.